@@ -1,7 +1,3 @@
 module github.com/imacks/cowtransfer
 
-go 1.14
-
-require (
-    github.com/orcaman/concurrent-map v0.0.0-20190826125027-8c72a8bb44f6
-)
+go 1.16
